+++ NINA/Model/MyCamera/ASICamera.cs
@@ -687,9 +687,9 @@ namespace NINA.Model.MyCamera {
             }
         }
 
-        public ArrayList Gains {
+        public IList<int> Gains {
             get {
-                return new ArrayList();
+                return new List<int>();
             }
         }