+++ NINA/Model/MyCamera/FLICamera.cs
@@ -345,7 +345,7 @@ namespace NINA.Model.MyCamera {
 
         public int GainMax => 0;
         public int GainMin => 0;
-        public ArrayList Gains => new ArrayList();
+        public IList<int> Gains => new List<int>();
         public bool HasBattery => false;
         public bool HasDewHeater => false;
         public bool HasSetupDialog => Connected ? true : false;