+++ NINA/Model/MyCamera/OmegonCamera.cs
@@ -432,9 +432,9 @@ namespace NINA.Model.MyCamera {
             }
         }
 
-        public ArrayList Gains {
+        public IList<int> Gains {
             get {
-                return new ArrayList();
+                return new List<int>();
             }
         }