+++ NINA/Model/MyCamera/Simulator/SimulatorCamera.cs
@@ -291,9 +291,9 @@ namespace NINA.Model.MyCamera.Simulator {
             }
         }
 
-        public ArrayList Gains {
+        public IList<int> Gains {
             get {
-                return null;
+                return new List<int>();
             }
         }